+++ src/config.ts
@@ -1,1 +1,0 @@
-export const FOLDER_NAME_SEGMENTS = "1_segments"